+++ __tests__/cases/server/server.js
@@ -40,6 +40,7 @@ const CONFIG_FIELDS = [
 	'cookies',
 	'createMissingComponents',
 	'generatedComponentsExtension',
+	'sourceStylesDir',
 ];
 
 describe('Server instance', () => {