+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Features
 - Application has `log[Info|Warning|Error]` methods to prevent using console.
 - Server has `getConfig` method to property access to the server configuration.
+- Option to change location of source styles that are merged into the application styles.
 ### Updates
 - Creating subdirectories if needed while creating components.